Moving the redactor function to its own file
This makes it more disconnected from the command line runner so we can use it to redact in other areas.

Updated the Docker service to use the redaction method to clean sanitize the buildArgs before logging
